requirements.txt
asttokens==2.4.1
attrs==23.2.0
autoattack @ git+https://github.com/fra31/auto-attack@a39220048b3c9f2cca9a4d3a54604793c68eca7e
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1717618050233/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.8.0
executing==2.0.1
fastjsonschema==2.20.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fsspec==2024.6.1
idna @ file:///croot/idna_1714398848350/work
ipython==8.12.3
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx @ file:///croot/networkx_1717597493534/work
numpy==2.0.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1714398848491/work
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
pyzmq==26.0.3
referencing==0.35.1
requests @ file:///croot/requests_1716902831423/work
rich==13.7.1
rpds-py==0.18.1
scipy==1.14.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
timm==0.6.5
tinycss2==1.3.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yarg==0.1.9